requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
bleach==2.1.3
certifi==2018.8.13
chardet==3.0.4
cycler==0.10.0
decorator
deepdiff==3.3.0
defusedxml==0.5.0
entrypoints==0.2.3
future==0.16.0
graphviz==0.8.4
h5py
html5lib==1.0.1
idna==2.7
igraph
ipykernel==4.8.2
ipython
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi
jgraph==0.2.1
joblib==0.13.2
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver
matplotlib
mistune==0.8.3
more-itertools==4.3.0
nbconvert==5.3.1
nbformat==4.4.0
networkx
numpy<2
of-adapt-vqe==0
openfermion
openfermionpsi4
pandocfilters==1.4.2
parso
pathlib2==2.3.3
pexpect
pickleshare==0.7.4
Pint==0.8.1
pluggy
prometheus-client
prompt-toolkit
ptyprocess
PubChemPy==1.0.4
py==1.5.4
pybind11
Pygments
pyparsing
pyscf
pytest
python-dateutil
pytz
#pyzmq==17.1.2
qtconsole==4.4.0
requests
scipy
Send2Trash==1.5.0
simplegeneric==0.8.1
six
terminado
testpath==0.3.1
tornado
traitlets
wcwidth
webencodings==0.5.1
widgetsnbextension==3.4.0
